conditional sdsl support and CI checking!

+++ src/fmindex-collection/occtable/GenericInterleaved.h
@@ -2,6 +2,7 @@
 
 #include "concepts.h"
 
+#include <algorithm>
 #include <array>
 #include <bitset>
 #include <cstdint>
@@ -22,7 +23,7 @@ namespace fmindex_collection {
 namespace occtable {
 namespace genericInterleaved {
 
-template <size_t TSigma, size_t TAlignment, typename block_t>
+template <uint64_t TSigma, uint64_t TAlignment, typename block_t>
 struct Bitvector {
     struct alignas(TAlignment) Block {
         std::array<block_t, TSigma> blocks{};
@@ -33,15 +34,15 @@ struct Bitvector {
 //            __builtin_prefetch((const void*)(&bits), 0, 0);
         }
 
-        uint64_t rank(size_t idx, size_t symb) const {
+        uint64_t rank(uint64_t idx, uint64_t symb) const {
             auto bitset = std::bitset<64>(bits[symb] << (63-idx));
             return blocks[symb] + bitset.count();
         }
 
-        uint64_t prefix_rank(size_t idx, size_t symb) const {
+        uint64_t prefix_rank(uint64_t idx, uint64_t symb) const {
             uint64_t b = {};
             uint64_t block = 0;
-            for (size_t i{0}; i <= symb; ++i) {
+            for (uint64_t i{0}; i <= symb; ++i) {
                 b = b | bits[i];
                 block += blocks[i];
             }
@@ -49,9 +50,9 @@ struct Bitvector {
             return block + bitset.count();
         }
 
-        size_t symbol(size_t idx) const {
+        uint64_t symbol(uint64_t idx) const {
             auto bit = (1ul << idx);
-            for (size_t symb{0}; symb < TSigma-1; ++symb) {
+            for (uint64_t symb{0}; symb < TSigma-1; ++symb) {
                 if (bits[symb] & bit) {
                     return symb;
                 }
@@ -65,14 +66,14 @@ struct Bitvector {
         }
     };
 
-    static constexpr size_t block_size = sizeof(block_t) * 8;
+    static constexpr uint64_t block_size = sizeof(block_t) * 8;
 
     std::vector<Block> blocks;
     std::vector<std::array<uint64_t, TSigma>> superBlocks;
     std::array<uint64_t, TSigma+1> C;
 
     template <typename CB>
-    Bitvector(size_t length, CB cb) {
+    Bitvector(uint64_t length, CB cb) {
         blocks.reserve(length/64+2);
 
         blocks.emplace_back();
@@ -81,7 +82,7 @@ struct Bitvector {
         std::array<uint64_t, TSigma> sblock_acc{0};
         std::array<block_t, TSigma> block_acc{0};
 
-        for (size_t size{1}; size <= length; ++size) {
+        for (uint64_t size{1}; size <= length; ++size) {
             if (size % (1ul<<block_size) == 0) { // new super block + new block
                 superBlocks.emplace_back(sblock_acc);
                 blocks.emplace_back();
@@ -102,7 +103,7 @@ struct Bitvector {
         }
 
         C[0] = 0;
-        for (size_t i{0}; i < TSigma; ++i) {
+        for (uint64_t i{0}; i < TSigma; ++i) {
             C[i+1] = sblock_acc[i] + C[i];
         }
     }
@@ -110,30 +111,30 @@ struct Bitvector {
     Bitvector(cereal_tag) {}
 
 
-    size_t memoryUsage() const {
+    uint64_t memoryUsage() const {
         return blocks.size() * sizeof(blocks.back())
             + superBlocks.size() * sizeof(superBlocks.back())
             + sizeof(C);
     }
 
-    void prefetch(size_t idx) const {
+    void prefetch(uint64_t idx) const {
         auto blockId      = idx >>  6;
         blocks[blockId].prefetch();
     }
 
-    uint64_t rank(uint64_t idx, size_t symb) const {
+    uint64_t rank(uint64_t idx, uint64_t symb) const {
         auto blockId      = idx >>  6;
         auto superBlockId = idx >> block_size;
         auto bitId        = idx &  63;
         return blocks[blockId].rank(bitId, symb) + superBlocks[superBlockId][symb] + C[symb];
     }
 
-    uint64_t prefix_rank(uint64_t idx, size_t symb) const {
+    uint64_t prefix_rank(uint64_t idx, uint64_t symb) const {
         auto blockId      = idx >>  6;
         auto superBlockId = idx >> block_size;
         auto bitId        = idx &  63;
         uint64_t a={};
-        for (size_t i{0}; i<= symb; ++i) {
+        for (uint64_t i{0}; i<= symb; ++i) {
             a += superBlocks[superBlockId][i];
         }
         return blocks[blockId].prefix_rank(bitId, symb) + a;
@@ -145,7 +146,7 @@ struct Bitvector {
         auto superBlockId = idx >> block_size;
         auto bitId        = idx &  63;
         auto res = std::array<uint64_t, TSigma>{};
-        for (size_t symb{0}; symb < TSigma; ++symb) {
+        for (uint64_t symb{0}; symb < TSigma; ++symb) {
             res[symb] = blocks[blockId].rank(bitId, symb) + superBlocks[superBlockId][symb] + C[symb];
         }
         return res;
@@ -159,25 +160,25 @@ struct Bitvector {
         auto rs  = std::array<uint64_t, TSigma>{};
         auto prs = std::array<uint64_t, TSigma>{};
 
-        for (size_t symb{0}; symb < TSigma; ++symb) {
+        for (uint64_t symb{0}; symb < TSigma; ++symb) {
             rs[symb]  = blocks[blockId].rank(bitId, symb);
         }
         rs[0] += superBlocks[superBlockId][0];
-        for (size_t symb{1}; symb < TSigma; ++symb) {
+        for (uint64_t symb{1}; symb < TSigma; ++symb) {
             rs[symb] += superBlocks[superBlockId][symb];
         }
         prs[0] = rs[0];
-        for (size_t symb{1}; symb < TSigma; ++symb) {
+        for (uint64_t symb{1}; symb < TSigma; ++symb) {
             prs[symb]= prs[symb-1] + rs[symb];
         }
 
-        for (size_t symb{0}; symb < TSigma; ++symb) {
+        for (uint64_t symb{0}; symb < TSigma; ++symb) {
             rs[symb] += C[symb];
         }
         return {rs, prs};
     }
 
-    size_t symbol(uint64_t idx) const {
+    uint64_t symbol(uint64_t idx) const {
         idx += 1;
         auto blockId      = idx >>  6;
         auto bitId        = idx &  63;
@@ -191,25 +192,25 @@ struct Bitvector {
 };
 
 
-template <size_t TSigma, size_t TAlignment, typename block_t>
+template <uint64_t TSigma, uint64_t TAlignment, typename block_t>
 struct OccTable {
     using TLengthType = uint64_t;
-    static constexpr size_t Sigma = TSigma;
+    static constexpr uint64_t Sigma = TSigma;
 
     Bitvector<Sigma, TAlignment, block_t> bitvector;
 
-    static size_t expectedMemoryUsage(size_t length) {
+    static uint64_t expectedMemoryUsage(uint64_t length) {
         using Block = typename Bitvector<TSigma, TAlignment, block_t>::Block;
         auto blockSize = std::max(alignof(Block), sizeof(Block));
 
-        size_t C           = sizeof(uint64_t) * (Sigma+1);
-        size_t blocks      = blockSize        * (length+1) / 64;
-        size_t superblocks = sizeof(uint64_t) * (length+1) / (1ul << (sizeof(block_t) * 8));
+        uint64_t C           = sizeof(uint64_t) * (Sigma+1);
+        uint64_t blocks      = blockSize        * (length+1) / 64;
+        uint64_t superblocks = sizeof(uint64_t) * (length+1) / (1ul << (sizeof(block_t) * 8));
         return C + blocks + superblocks;
     }
 
     OccTable(std::vector<uint8_t> const& _bwt)
-        : bitvector(_bwt.size(), [&](size_t i) -> uint8_t {
+        : bitvector(_bwt.size(), [&](uint64_t i) -> uint8_t {
             return _bwt[i];
         })
     {}
@@ -218,7 +219,7 @@ struct OccTable {
         : bitvector(cereal_tag{})
     {}
 
-    size_t memoryUsage() const {
+    uint64_t memoryUsage() const {
         return bitvector.memoryUsage() + sizeof(OccTable);
     }
 
@@ -230,15 +231,15 @@ struct OccTable {
         bitvector.prefetch(idx);
     }
 
-    uint64_t rank(uint64_t idx, size_t symb) const {
+    uint64_t rank(uint64_t idx, uint64_t symb) const {
         return bitvector.rank(idx, symb);
     }
 
-    uint64_t prefix_rank(uint64_t idx, size_t symb) const {
+    uint64_t prefix_rank(uint64_t idx, uint64_t symb) const {
         return bitvector.prefix_rank(idx, symb);
     }
 
-    size_t symbol(uint64_t idx) const {
+    uint64_t symbol(uint64_t idx) const {
         return bitvector.symbol(idx);
     }
